1. Run Image server using bundle. 2. Use cors package instead of our own cors handling in image server where express is used.
