+++ build_image_server.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+import os 
+
+os.system('node utils/build_image_server.js')
+print("Completed")
