--- licit/server/image/index.js
@@ -1,6 +1,0 @@
-
-function main(): void {
-  console.log('Image Server running...');
-}
-
-window.onload = main;
