+++ run_collab_server.py
@@ -3,12 +3,10 @@
 
 import os
 import platform
-import logging
-import socket    
-from subprocess import check_output
+import socket
 
-# [FS-NK][03-MAR-2020]
-# IRAD-892 Correct licit windows build
+# [FS] IRAD-892 2020-03-03
+# Correct licit windows build
 # gracefully handling commands WRT OS.
 hostname = socket.gethostname()    
 IPAddr = socket.gethostbyname(hostname)  
@@ -16,6 +14,7 @@ port = '3002'
 
 if platform.system()=="Windows": cmd = 'node servers/collab/run_licit_collab_server.bundle.js ' + ' PORT=' + port + ' IP='+ IPAddr
 else: cmd = 'PORT=' + port + ' IP=' + IPAddr + ' node servers/collab/run_licit_collab_server.bundle.js'
+
 print('=' * 80)
 print(cmd)
 print('=' * 80)
