+++ run_image_server.py
@@ -3,23 +3,18 @@
 
 import os
 import platform
-import logging
-import socket    
-from subprocess import check_output
+import socket
 
-# [FS][03-MAR-2020]
-# IRAD-892 Correct licit windows build
+# [FS] IRAD-991 2020-07-17
 # gracefully handling commands WRT OS.
 hostname = socket.gethostname()    
 IPAddr = socket.gethostbyname(hostname)  
 port = '3004'
 
-if platform.system()=="Windows": cmd = 'node utils/build_image_server.js ' + ' PORT=' + port + ' IP='+ IPAddr
-else: cmd = 'PORT=' + port + ' IP=' + IPAddr + ' node utils/build_image_server.js'
+if platform.system()=="Windows": cmd = 'node servers/image/run_image_server.bundle.js ' + ' PORT=' + port + ' IP='+ IPAddr
+else: cmd = 'PORT=' + port + ' IP=' + IPAddr + ' node servers/image/run_image_server.bundle.js'
 
 print('=' * 80)
 print(cmd)
-print('=' * 80)
-print('run image server at http://' + IPAddr + ':' + port)
 print('=' * 80)
 os.system(cmd)
